pyproject.toml
[tool.poetry]
name = "simple_diffusion"
version = "0.1.0"
description = "Simple diffusion model"
authors = ["Maxim Grigorev <gignm01@gmail.com>"]
maintainers = ["Maxim Grigorev <gignm01@gmail.com>"]
readme = "README.md"
repository = "https://github.com/Jserax/simple_diffusion"
keywords = ["data science", "template"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "3.10.12"
torch = "^2.1.0"
pillow = "^10.1.0"
pandas = "^2.1.3"
transformers = "^4.35.2"
tqdm = "^4.66.1"
torchvision = "^0.16.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
pytest = "^7.4.3"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
src_paths = ["simple_diffusion"]
profile = "black"
line_length = 88
lines_after_imports = 2

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
